requirements.txt
ddddocr
playwright
pyyaml